+++ test/host/mocks/esp_err.h
@@ -15,6 +15,7 @@ typedef int esp_err_t;
 #define ESP_ERR_INVALID_SIZE    0x104
 #define ESP_ERR_NOT_FOUND       0x105
 #define ESP_ERR_NOT_SUPPORTED   0x106
+#define ESP_ERR_INVALID_RESPONSE 0x108
 #define ESP_ERR_TIMEOUT         0x107
 
 #define ESP_ERROR_CHECK(x) do {                                     \